[Feature] Add fillColorApplied to Character and Paragraph Style
