+++ src/tests/controllers/ParagraphStyleController.test.ts
@@ -1,6 +1,6 @@
-import { ColorType } from '../../../types/ColorStyleTypes';
+import { ColorType, ColorUsageType } from '../../../types/ColorStyleTypes';
 import { ParagraphStyleController } from '../../controllers/ParagraphStyleController';
-import { ColorUsageType, ParagraphStyleUpdate } from '../../../types/ParagraphStyleTypes';
+import { ParagraphStyleUpdate } from '../../../types/ParagraphStyleTypes';
 import { Alignment, Case, Scripting } from '../../../types/TextStyleTypes';
 import MockEditorAPI from '../__mocks__/MockEditorAPI';
