+++ types/ParagraphStyleTypes.ts
@@ -1,56 +1,5 @@
 import { Alignment, Case, Scripting } from './TextStyleTypes';
-import { Color, ColorUpdate } from './ColorStyleTypes';
-
-export type ColorUsageUpdate = {
-    /**
-     * Reference to the stylekit color, can only be used if it is a stylekit color.
-     */
-    colorId?: string;
-    /**
-     * The local color object, can only be used if it is a local color.
-     */
-    color?: ColorUpdate;
-    usageType: ColorUsageType;
-    /**
-     * Opacity of the color (0-100) where 0 = fully transparent.
-     */
-    opacity?: number;
-    /**
-     * Indicates if the color will be applied or not.
-     */
-    isApplied?: boolean;
-};
-
-export type ColorUsage = {
-    /**
-     * Reference to the stylekit color, can only be used if it is a stylekit color.
-     */
-    colorId?: string;
-    /**
-     * The local color object, can only be used if it is a local color.
-     */
-    color?: Color;
-    usageType: ColorUsageType;
-    /**
-     * Opacity of the color (0-100) where 0 = fully transparent.
-     */
-    opacity?: number;
-    /**
-     * Indicates if the color will be applied or not.
-     */
-    isApplied?: boolean;
-};
-
-export enum ColorUsageType {
-    /**
-     * The color is defined locally on the object
-     */
-    local = 'local',
-    /**
-     * THe color is referencing a stylekit color
-     */
-    stylekit = 'stylekit',
-}
+import {Color, ColorUpdate, ColorUsage, ColorUsageUpdate} from './ColorStyleTypes';
 
 export type ParagraphStyle = {
     id: string;
@@ -77,6 +26,7 @@ export type ParagraphStyle = {
     textAlignLast: Alignment;
     textOverprint: boolean;
     color: ColorUsage;
+    fillColorApplied: boolean;
     underline: boolean;
     lineThrough: boolean;
 };
